+++ packages/server-wallet/src/config/defaults.ts
@@ -1,7 +1,4 @@
-import {utils} from 'ethers';
 import _ from 'lodash';
-
-import {IncomingEngineConfigV2} from '../engine/engine';
 
 import {
   ChainServiceConfiguration,
@@ -12,8 +9,6 @@ import {
   OptionalWalletConfig,
   WalletConfig,
 } from './types';
-
-import {extractDBConfigFromWalletConfig} from '.';
 
 export const defaultDatabaseConfiguration: OptionalDatabaseConfiguration & {
   connection: {host: string; port: number; user: string};
@@ -67,19 +62,6 @@ type HasDatabaseConnectionConfigObject = {
 export const defaultTestNetworkConfiguration: NetworkConfiguration = {
   // eslint-disable-next-line no-process-env
   chainNetworkID: parseInt(process.env.CHAIN_NETWORK_ID ?? '0'),
-};
-export const defaultTestEngineConfig = (
-  partialConfig?: Partial<IncomingEngineConfigV2>
-): IncomingEngineConfigV2 => {
-  const defaultEngineConfig: IncomingEngineConfigV2 = {
-    skipEvmValidation: true,
-    metrics: {timingMetrics: false},
-    workerThreadAmount: 0,
-    // eslint-disable-next-line no-process-env
-    chainNetworkID: utils.hexlify(parseInt(process.env.CHAIN_NETWORK_ID ?? '0')),
-    dbConfig: extractDBConfigFromWalletConfig(defaultTestWalletConfig()),
-  };
-  return _.assign({}, defaultEngineConfig, partialConfig);
 };
 
 export const defaultTestWalletConfig = (
