+++ packages/server-wallet/src/engine/__test__/integration/create-ledger-channel.test.ts
@@ -2,10 +2,10 @@ import {constants} from 'ethers';
 import {NULL_APP_DATA} from '@statechannels/wallet-core';
 
 import {Channel} from '../../../models/channel';
-import {Engine} from '../..';
+import {defaultTestEngineConfig, Engine} from '../..';
 import {createChannelArgs} from '../fixtures/create-channel';
 import {seedAlicesSigningWallet} from '../../../db/seeds/1_signing_wallet_seeds';
-import {defaultTestWalletConfig, defaultTestEngineConfig} from '../../../config';
+import {defaultTestWalletConfig} from '../../../config';
 import {DBAdmin} from '../../../db-admin/db-admin';
 import {createLogger} from '../../../logger';
